LPS-47201 These test utilities are used during application context startup during tests, so it doesn't make sense having them under portal-test. Moving them to portal-impl/test/integration
