+++ portal-impl/test/integration/com/liferay/portal/security/pacl/PACLIntegrationJUnitTestRunner.java
@@ -15,7 +15,7 @@
 package com.liferay.portal.security.pacl;
 
 import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.service.test.ServiceTestUtil;
+import com.liferay.portal.service.ServiceTestUtil;
 import com.liferay.portal.test.LiferayIntegrationJUnitTestRunner;
 
 import java.net.MalformedURLException;
